fixed flags

+++ Assets/Koodi/Slowable.cs
@@ -2,6 +2,6 @@
     public interface Slowable {
         public void slowDown();
         public void speedUp();
-        public void addSpeed(int percent);
+        public void addSpeed(float percent);
     }
 }
